added new changes

+++ data-structures/linked-list/linked-list.js
@@ -64,12 +64,18 @@ LinkedList.prototype.contains = function (target) {
 
 LinkedList.prototype.addToHead = function (value) {
   var newNode = new Node(value, this.head, null);
-  if (this.head) {
-    this.head.prev = newNode;
-  } else {
-    this.tail = newNode;
+  if(value){
+    if (this.head) {
+      this.head.prev = newNode;
+    } else {
+      this.tail = newNode;
+    }
+    this.head = newNode;
   }
-  this.head = newNode;
+  else{
+    resultText.innerHTML = `Error! cannot add null value, enter a value`;
+  }
+
 };
 
 //method to delete node from linked list given a value
@@ -173,21 +179,20 @@ LinkedList.prototype.printList = function () {
   }
   //change the innerHTML of resultText the value of this.head and this.tail
   if(this.head && this.tail){
-  resultText.innerHTML = `{This is the Head: ${this.head.value}} 
-  {This is the Tail: ${this.tail.value}}`;
+  resultText.innerHTML = `The Value conatined at the head is: ${this.head.value} & 
+  The Value conatined at the Tail is: ${this.tail.value}`;
   }
   else if(this.head === null){
-    resultText.innerHTML = `{This is the Head: ${this.head}} 
-    {This is the Tail: ${this.tail}}`;
+    resultText.innerHTML = `All Values have been removed`;
   }
-  else if(this.tail === null){
-    resultText.innerHTML = `{This is the Head: ${this.head.value}}
-    {This is the Tail: ${this.tail}}`;
+  else if(!this.tail){
+    resultText.innerHTML = `The Value conatined at the head is: ${this.head.value} & 
+    The Value conatined at the Tail is only tail: ${this.tail.value}`;
   }
 
   else if(this.head === null && this.tail === null){
-    resultText.innerHTML = `{This is the Head: ${this.head}}
-    {This is the Tail: ${this.tail}}`;
+    resultText.innerHTML = `The Value conatined at the head is: ${this.head.value} & 
+    The Value conatined at the Tail is yeaaa: ${this.tail.value}`;
   }
   //change the innerHTML of resEle to result
   resEle.innerHTML = result;
@@ -202,24 +207,31 @@ list.addToHead(3);
 //insert button event listener executes addToTail method when the insert button is clicked
 //adds the value in the input field to the tail (end) of the linked list
 insertButton.addEventListener("click", function () {
-  //call addToTail method
-  list.addToTail(insertValue.value);
-
-  //print list
+  if(insertHeadValue.value){
+    list.addToTail(insertValue.value);     //print list
   list.printList();
+  }
+  else{
+    resultText.innerHTML = `Error! cannot add null value, enter a value`;
+  }
 
-  //log the list
-  console.log("This is the list after insert tail", list);
 });
 
 //insert head button event listener executes addToHead method when the insert head button is clicked
 //inserts the element at the beginning of the list
 insertHeadButton.addEventListener("click", function () {
   //call the addToHead method
-  list.addToHead(insertHeadValue.value);
-
-  //print list
+  if(insertHeadValue.value){
+    list.addToHead(insertHeadValue.value);
+     //print list
   list.printList();
+  }
+  else{
+    resultText.innerHTML = `Error! cannot add null value, enter a value`;
+  }
+ 
+
+ 
 
   //log the list
   console.log("This is the list after insert head", list);
@@ -229,6 +241,7 @@ insertHeadButton.addEventListener("click", function () {
 //removes the first element from the list
 removeHeadButton.addEventListener("click", function () {
   //calling the removeHead method
+  
   list.removeHead();
 
   //print list
@@ -266,38 +279,18 @@ removeTailButton.addEventListener("click", function () {
 //removes the node that matches the value provided
 deleteButton.addEventListener("click", function () {
   //calling the removeNode method
-  list.deleteNode(Number(deleteValue.value));
+  if(deleteValue.value){
+    list.deleteNode(Number(deleteValue.value));
   console.log("This is the list after delete node", list.deleteNode(deleteValue.value));
 
   //print list
   list.printList();
+  }
+  else{
+    resultText.innerHTML = `Error! cannot delete, please enter a value`;
+  }
+  
 
   //log the list
   console.log("This is the list after remove node", list);
 });
-
-
-//Pseudocode
-const container = document.getElementById("array");
-let currentArray = [];
-const setCountValue = document.getElementById("countValue");
-let codeExplaination = document.getElementById('codeExplaination');
-const header = document.querySelector(".header");
-let pseudo = ['1. for i ← 0, a.length − 1 do', 'if (a[i] == e) then', '3 return true', 'end if', 'end for', 'return false'];
-
-let LinkedListpseudo = ['AddToTail, if tail.next = newNode, else head = newNode', 
-'RemoveHead, if head.prev = null, else tail = null', 
-'AddToHead, if head.prev = newNode, else tail = newNode', 
-'DeleteNode, while node, if node.value === value, node.prev.next = node.next, else head = node.next, if node.next, node.next.prev = node.prev, else tail = node.prev', 
-'RemoveTail, if not tail RETURN, if tail.next = null, else head = null', 
-];
-
-
-setCountValue.innerHTML +=`
-  <li id='setCountValue1' class = 'listCss'>${countingPsuedo[0]}</li>
-  <li id='setCountValue2'  class = 'listCss'>${countingPsuedo[1]}</li>
-  <li id='setCountValue3'  class = 'listCss'>${countingPsuedo[2]}</li>
-  <li id='setCountValue4'  class = 'listCss'>${countingPsuedo[3]}</li>
-  <li id='setCountValue5'  class = 'listCss'>${countingPsuedo[4]}</li>
-  <li id='setCountValue6'  class = 'listCss'>${countingPsuedo[5]}</li>
-  `
